Implement a postgres database connection pool
Connection information and pool parameters are parsed from config.toml,
and a connection pool is created and managed by deadpool. We communicate
with postgres via the tokio_postgres crate.

I thaught about using an ORM but A: diesel isn't the most async-friendly
thing in the world and B: I think it's overkill for our usecase given
how many extremely similar queries we'll be doing.

I also considered SQLX, but the added dependencies and the fact that I
want to use postgres-specific features, meaning the cross-database
compatibility is useless for us, turned me away