Factor tx/rx pattern out into util::ReadWrite<R, W> struct
For both sockets and channels, we need to have two "ends" for borrowing
reasons, each of which need to be inside a *separately lockable* Mutex
for performance reasons. This could also be extended to many other
objects in the future (E.G. any kind of IO, mpsc channels, etc).

The ReadWrite<R, W> struct abstracts this pattern, and provides some
convenience constructers and methods for creating and working with these
kinds of connected pairs. Each end is automatically wrapped in a Mutex.
In a sense, this works like an RwLock, except .read() and .write()
access separate objects.

This has been implemented in the Connection struct, to great success!